added file-record microservice, abstract-role guard

+++ apps/profiles/src/profiles.service.ts
@@ -1,8 +1,11 @@
 import {
   CREATE_USER,
+  DELETE_FILE,
   DELETE_USER,
   GET_TOKEN,
   GET_USER_BY_EMAIL,
+  RECORD_FILE,
+  UPDATE_FILE,
   UPDATE_USER,
 } from '@app/shared';
 import { CreateUserProfileDto } from '@app/shared/dto/create-user-profile.dto';
@@ -20,6 +23,7 @@ export class ProfilesService {
     @InjectRepository(Profile)
     private readonly profileRepository: Repository<Profile>,
     @Inject('AUTH') private authService: ClientProxy,
+    @Inject('FILES-RECORD') private fileRecordService: ClientProxy,
     private fileService: FilesService,
   ) {}
 
@@ -56,9 +60,23 @@ export class ProfilesService {
         ...userProfileDto,
         userId: newUser.id,
         photo: namePhoto,
-        creationDate: new Date(),
       });
       await this.profileRepository.save(profile);
+      const createdProfile = await this.profileRepository.findOneBy({
+        userId: newUser.id,
+      });
+      if (namePhoto) {
+        await firstValueFrom(
+          this.fileRecordService.send(
+            { cmd: RECORD_FILE },
+            {
+              essenceId: createdProfile.id,
+              essenceTable: 'profiles',
+              fileName: namePhoto,
+            },
+          ),
+        );
+      }
       return await firstValueFrom(
         this.authService.send({ cmd: GET_TOKEN }, newUser),
       );
@@ -92,14 +110,36 @@ export class ProfilesService {
         }
       }
       if (namePhoto && profile.photo) {
-        await this.fileService.deleteFile(profile.photo);
-        await this.profileRepository.save({
-          ...profile,
-          ...userProfileDto,
-          photo: namePhoto,
-        });
+        if (profile.photo) {
+          await this.fileService.deleteFile(profile.photo);
+        } else {
+          await this.profileRepository.save({
+            ...profile,
+            ...userProfileDto,
+            photo: namePhoto,
+          });
+          await firstValueFrom(
+            this.fileRecordService.send(
+              { cmd: UPDATE_FILE },
+              {
+                essenceId: profileId,
+                fileName: namePhoto,
+              },
+            ),
+          );
+        }
       } else {
         await this.profileRepository.save({ ...profile, ...userProfileDto });
+        await firstValueFrom(
+          this.fileRecordService.send(
+            { cmd: RECORD_FILE },
+            {
+              essenceId: profileId,
+              essenceTable: 'profiles',
+              fileName: namePhoto,
+            },
+          ),
+        );
       }
 
       return await this.profileRepository.findOneBy({ id: profileId });
@@ -127,9 +167,17 @@ export class ProfilesService {
       if (user.status === 'error') {
         return user;
       }
-      await this.fileService.deleteFile(profile.photo);
+      if (profile.photo) {
+        await this.fileService.deleteFile(profile.photo);
+        await firstValueFrom(
+          this.fileRecordService.send({ cmd: DELETE_FILE }, profileId),
+        );
+      }
       await this.profileRepository.remove(profile);
-      return 'Success';
+      return {
+        status: 'Success',
+        message: 'Профиль и пользователь успешно удалены',
+      };
     } catch (error) {
       return {
         status: 'error',
